back-tif/pyproject.toml
[tool.fastapi]
host = "0.0.0.0"
port = 8000
reload = true

[project]
name = "tif-backend"
version = "1.0.0"
description = "Backend para detección de fechas de vencimiento"
dependencies = [
    # FastAPI y servidor web
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    
    # Base de datos
    "pymongo>=4.6.0",
    "motor>=3.3.2",
    
    # Utilidades
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "aiohttp>=3.8.0",  # Para comunicación con servicio FCOS
    
    # Pydantic para validación de datos
    "pydantic>=2.5.0",
    
    # Dependencias para el modelo DAN (OCR)
    "torch>=2.0.0",
    "torchvision>=0.15.0", 
    "torchaudio>=2.0.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "editdistance>=1.0.0",
    "Colour>=0.1.5",
    "LMDB>=1.4.0",
    
    # Dependencias para FCOS/AdelaiDet (detección de objetos)
    "detectron2",
    "fvcore",
    "pycocotools",
    "PyYAML",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta" 